+++ apps/storybook/.storybook/main.cjs
@@ -1,8 +1,12 @@
 const svgr = require("vite-plugin-svgr").default;
 const { mergeConfig } = require("vite");
 module.exports = {
-  stories: [ "../src/**/*.stories.@(js|jsx|ts|tsx)"],
+  stories: [
+    "../src/**/*.stories.@(js|jsx|ts|tsx)",
+    "../src/**/*.mdx",
+  ],
   addons: [
+    "@storybook/addon-docs",
     "@storybook/addon-links",
     "@storybook/addon-essentials",
     "@storybook/addon-interactions",
@@ -10,9 +14,6 @@ module.exports = {
   framework: {
     name: "@storybook/react-vite",
     options: {},
-  },
-  docs: {
-    autodocs: "tag",
   },
   async viteFinal(config) {
     // Merge custom configuration into the default config