+++ packages/color-tag/index.ts
@@ -1,3 +1,8 @@
 import "@life-uikit/global-css/tailwind-css/input.css";
 
 export { ColorTag } from "./src";
+export { ColorPikcer } from "./src/color-picker";
+export { TagCreate } from "./src/tag-create";
+export { TagManagement } from "./src/tag-management";
+export { ColorTag as ColorTagBase } from "./src/color-tag";
+export { TagItems } from "./src/tag-items";